requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==3.2
ifaddr==0.1.7
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
soco==0.23.3
spotipy==2.19.0
urllib3==1.26.6
xmltodict==0.12.0
Flask==2.0.1
mfrc522==0.0.7
RPi.GPIO==0.7.0